Issue 2: Non Owner Image Access resolved for Delete

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -148,9 +148,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession httpSession, Model model) {
+        Image image = imageService.getImage(imageId);
+        if(imageService.isOwner(image,httpSession)){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else{
+            String error = "Only the owner of the image can delete the image";
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            model.addAttribute("deleteError",error);
+            return "images/image";
+        }
     }
